requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.1.1
ctranslate2==4.0.0
cycler==0.12.1
dataclasses-json==0.6.4
decorator==4.4.2
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
faster-whisper==1.0.1
filelock==3.13.1
flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==6.0.1
HyperPyYAML==1.2.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.2
importlib-resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
julius==0.2.7
kiwisolver==1.4.5
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.1
langsmith==0.1.23
lazy-loader==0.3
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.10.1
llama-index==0.10.19
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.9
llama-index-core==0.10.19
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.4
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.9
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.11
llama-index-readers-llama-parse==0.1.3
llama-parse==0.3.9
llamaindex-py-client==0.1.13
llvmlite==0.41.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.5
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.16.3
openai==1.13.3
optuna==3.5.0
orjson==3.9.15
packaging==23.2
pandas==2.0.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
portalocker==2.8.2
primePy==1.3
proglog==0.1.10
protobuf==4.25.3
psycopg2-binary==2.9.9
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.6.3
pydantic-core==2.16.3
pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.8.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.2
sentence-transformers==2.5.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.28
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.1
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typer==0.9.0
typing-extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
